+++ Modules/ItemList/SearchBox.lua
@@ -6,20 +6,22 @@ NS.ItemListSearchBox = SearchBox
 
 local SEARCH_BOX_TEMPLATE = "SearchBoxNineSliceTemplate"
 local SEARCH_BOX_WIDTH = 320
-local SEARCH_BOX_HEIGHT = 28
-local SEARCH_BOX_TEXT_SIZE = 18
+local SEARCH_BOX_TEXT_SIZE = 13
 local SEARCH_BOX_FONT_FLAGS = ""
-local SEARCH_ICON_LEFT_OFFSET = 4
-local SEARCH_ICON_Y_OFFSET = 0
-
-local function GetPrimaryFont()
-    return NS.Media.GetPrimaryFont()
-end
+local SEARCH_ICON_LEFT_OFFSET = 7
+local SEARCH_ICON_Y_OFFSET = 1
+local CLEAR_BUTTON_RIGHT_OFFSET = -4
+local CLEAR_BUTTON_Y_OFFSET = 1
+local SEARCH_TEXT_LEFT_INSET = 25
+local SEARCH_TEXT_RIGHT_INSET = 23
 
 function SearchBox.Create(parent, list)
     local searchBox = CreateFrame("EditBox", nil, parent, SEARCH_BOX_TEMPLATE)
-    searchBox:SetSize(SEARCH_BOX_WIDTH, SEARCH_BOX_HEIGHT)
-    searchBox:SetFont(GetPrimaryFont(), SEARCH_BOX_TEXT_SIZE, SEARCH_BOX_FONT_FLAGS)
+    searchBox:SetWidth(SEARCH_BOX_WIDTH)
+    searchBox.Background:SetAtlas(NS.Media.GetTextInputBackgroundAtlas(), false)
+    searchBox:SetFont(NS.Media.GetPrimaryFont(), SEARCH_BOX_TEXT_SIZE, SEARCH_BOX_FONT_FLAGS)
+    searchBox:SetJustifyV("MIDDLE")
+    searchBox:SetTextInsets(SEARCH_TEXT_LEFT_INSET, SEARCH_TEXT_RIGHT_INSET, 0, 0)
     searchBox.searchIcon:ClearAllPoints()
     searchBox.searchIcon:SetPoint(
         "LEFT",
@@ -28,19 +30,21 @@ function SearchBox.Create(parent, list)
         SEARCH_ICON_LEFT_OFFSET,
         SEARCH_ICON_Y_OFFSET
     )
-    if searchBox.Instructions then
-        searchBox.Instructions:SetFont(GetPrimaryFont(), SEARCH_BOX_TEXT_SIZE, SEARCH_BOX_FONT_FLAGS)
-    end
+    searchBox.clearButton:ClearAllPoints()
+    searchBox.clearButton:SetPoint(
+        "RIGHT",
+        searchBox,
+        "RIGHT",
+        CLEAR_BUTTON_RIGHT_OFFSET,
+        CLEAR_BUTTON_Y_OFFSET
+    )
+    searchBox.Instructions:SetFont(NS.Media.GetPrimaryFont(), SEARCH_BOX_TEXT_SIZE, SEARCH_BOX_FONT_FLAGS)
+    searchBox.Instructions:ClearAllPoints()
+    searchBox.Instructions:SetPoint("TOPLEFT", searchBox, "TOPLEFT", SEARCH_TEXT_LEFT_INSET, 0)
+    searchBox.Instructions:SetPoint("BOTTOMRIGHT", searchBox, "BOTTOMRIGHT", -SEARCH_TEXT_RIGHT_INSET, 0)
     searchBox:SetAutoFocus(false)
-    searchBox:SetScript("OnTextChanged", function(self)
-        if SearchBoxTemplate_OnTextChanged then
-            SearchBoxTemplate_OnTextChanged(self)
-        end
-
+    searchBox:HookScript("OnTextChanged", function(self)
         list:SetSearchText(self:GetText())
-    end)
-    searchBox:SetScript("OnEscapePressed", function(self)
-        self:ClearFocus()
     end)
 
     return searchBox